pyproject.toml
[project]
name = "msr"
version = "0.1.0"
description = "MSR (Multi Speaker Recognization) offline transcription service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "numpy>=1.26.0,<3.0.0",
  "librosa>=0.10.2,<1.0.0",
  "soundfile>=0.12.1,<1.0.0",
  "psutil>=5.9.8,<8.0.0",
  "setuptools>=80.0.0,<81.0.0",
  "webrtcvad>=2.0.10,<3.0.0",
  "nvidia-ml-py>=12.560.30,<13.0.0",
]

[project.optional-dependencies]
funasr = [
  "funasr>=1.2.0",
]
faster-whisper = [
  "faster-whisper>=1.1.0",
]
3d-speaker = [
  "speakerlab",
]
pyannote = [
  "pyannote.audio>=3.1.0",
]
gpu-runtime = [
  "funasr>=1.2.0",
  "faster-whisper>=1.1.0",
  "speakerlab",
  "pyannote.audio>=3.1.0",
]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[project.scripts]
msr-api = "msr.app.main:run"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/msr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
